oauth: Add silex controller for oauth routes
The controller has two routes jira-connect and jira-callback. An
application can start the authentication process with jira calling the
jira-connect route.

The jira-connect route will trigger the temporary token request and will
then redirect to the JIRA OAuth page. After the user has logged in and
allowed the application to access JIRA it will redirect back to the
jira-callback route.

The jira-callback route will process the OAuth data from JIRA and create
an access token, which can be used for making REST API request in the
future. If this is successful the user will be redirected to a page of
the original application.