Add note about minimum stability
Without setting the minimum stability to "dev" composer will not be able
to resolve the guzzlehttp/oauth-provider package.

We need the current dev-master of the oauth-provider as it contains a
fix that was made after the 0.2 release. This fix is required if a
private key file is used, which is the case with JIRA.

Once the guzzelhttp/oauth-subscriber has a new release, the dependency
should point to a specific release, removing the necessity to set the
minimum stability.

closes: #1