Set oauth-subscriber version to dev-master
Using the provider required setting the minimum-stability to dev, so the
oauth-subscriber package could be resolved. But setting this will also
pull the dev versions of all other packages.

By setting the version of the oauth-subscriber to dev-master, the
minimum-stability flag is no longer necessary and the oauth-subscriber
package will still be pulled in the correct version.